refactor: components moved to new directories
