+++ src/components/dialog/dialog-delete.js
