+++ src/pages/global-specific-settings.js
@@ -4,7 +4,7 @@ import { Button, ButtonStrip } from '@dhis2/ui-core'
 import i18n from '@dhis2/d2-i18n'
 import SettingsTable from '../components/settings-table/settings-table'
 import ProgramGlobalSettings from '../components/settings-table/program-global-settings'
-import DialogDelete from '../components/dialog-delete'
+import DialogDelete from '../components/dialog/dialog-delete'
 import DialogTable from '../components/dialog-table'
 import TableActions from '../components/table-actions'
 import DialogSaveData from '../components/dialog/dialog-save-data'
